+++ order-service/src/main/java/n11bootcamp_project_backend/order_service/enums/OrderStatus.java
@@ -1,0 +1,7 @@
+package n11bootcamp_project_backend.order_service.enums;
+
+public enum OrderStatus {
+    PENDING,
+    CONFIRMED,
+    CANCELLED
+}